# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inquiries:
    get:
      operationId: listInquiries
      tags: [inquiries]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Inquiry"
    post:
      operationId: createInquiry
      tags: [inquiries]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInquiryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inquiry"
  /inquiries/{id}:
    get:
      operationId: getInquiry
      tags: [inquiries]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inquiry"
    patch:
      operationId: updateInquiry
      tags: [inquiries]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInquiryInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inquiry"
    delete:
      operationId: deleteInquiry
      tags: [inquiries]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /inquiries/{id}/convert:
    post:
      operationId: convertInquiry
      tags: [inquiries]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /follow-ups:
    get:
      operationId: listFollowUps
      tags: [followups]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FollowUp"
    post:
      operationId: createFollowUp
      tags: [followups]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFollowUpInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowUp"
  /follow-ups/{id}:
    patch:
      operationId: updateFollowUp
      tags: [followups]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFollowUpInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowUp"
    delete:
      operationId: deleteFollowUp
      tags: [followups]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /bills:
    get:
      operationId: listBills
      tags: [bills]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bill"
    post:
      operationId: createBill
      tags: [bills]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBillInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
  /bills/{id}:
    patch:
      operationId: updateBill
      tags: [bills]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBillInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
...
components:
  schemas:
...
    Member:
      type: object
      required: [id, name, phone, status, plan, joinDate, expiryDate, gender, avatarUrl, balanceDue]
      properties:
        id: { type: string }
        name: { type: string }
        phone: { type: string }
        email: { type: ["string", "null"] }
        status: { type: string }
        plan: { type: string }
        joinDate: { type: string, format: date-time }
        expiryDate: { type: string, format: date-time }
        gender: { type: string }
        bloodGroup: { type: ["string", "null"] }
        dob: { type: ["string", "null"], format: date-time }
        address: { type: ["string", "null"] }
        avatarUrl: { type: string }
        balanceDue: { type: integer }
        medicalHistory: { type: ["string", "null"] }
        emergencyContact: { type: ["string", "null"] }
...
    Inquiry:
      type: object
      required: [id, name, phone, interest, source, status, inquiryDate, assignedTo]
      properties:
        id: { type: string }
        name: { type: string }
        phone: { type: string }
        email: { type: ["string", "null"] }
        interest: { type: string }
        source: { type: string }
        status: { type: string }
        inquiryDate: { type: string, format: date-time }
        assignedTo: { type: string }
        notes: { type: ["string", "null"] }
    CreateInquiryInput:
      type: object
      required: [name, phone, interest, source]
      properties:
        name: { type: string }
        phone: { type: string }
        email: { type: ["string", "null"] }
        interest: { type: string }
        source: { type: string }
        assignedTo: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
    UpdateInquiryInput:
      type: object
      properties:
        name: { type: string }
        phone: { type: string }
        email: { type: ["string", "null"] }
        interest: { type: string }
        source: { type: string }
        status: { type: string }
        assignedTo: { type: string }
        notes: { type: ["string", "null"] }
    FollowUp:
      type: object
      required: [id, memberId, memberName, type, priority, scheduledFor, status, lastComment, phone]
      properties:
        id: { type: string }
        memberId: { type: ["string", "null"] }
        memberName: { type: string }
        type: { type: string }
        priority: { type: string }
        scheduledFor: { type: string, format: date-time }
        status: { type: string }
        lastComment: { type: string }
        phone: { type: string }
    CreateFollowUpInput:
      type: object
      required: [memberName, type, priority, scheduledFor, phone]
      properties:
        memberId: { type: ["string", "null"] }
        memberName: { type: string }
        type: { type: string }
        priority: { type: string }
        scheduledFor: { type: string, format: date-time }
        lastComment: { type: ["string", "null"] }
        phone: { type: string }
    UpdateFollowUpInput:
      type: object
      properties:
        type: { type: string }
        priority: { type: string }
        scheduledFor: { type: string, format: date-time }
        status: { type: string }
        lastComment: { type: string }
...
    Bill:
      type: object
      required: [id, memberId, memberName, type, amount, paid, balance, status, createdAt]
      properties:
        id: { type: string }
        memberId: { type: string }
        memberName: { type: string }
        type: { type: string }
        items:
          type: array
          items:
            type: object
            required: [name, amount]
            properties:
              name: { type: string }
              amount: { type: integer }
        amount: { type: integer }
        paid: { type: integer }
        balance: { type: integer }
        status: { type: string }
        paymentMode: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    CreateBillInput:
      type: object
      required: [memberId, type, items]
      properties:
        memberId: { type: string }
        type: { type: string }
        items:
          type: array
          items:
            type: object
            required: [name, amount]
            properties:
              name: { type: string }
              amount: { type: integer }
        paid: { type: integer }
        paymentMode: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
    UpdateBillInput:
      type: object
      properties:
        paid: { type: integer }
        status: { type: string }
        paymentMode: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
...