# lib/api-zod/src/generated/types/bill.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BillItemsItem } from "./billItemsItem";

export interface Bill {
  id: string;
  memberId: string;
  memberName: string;
  type: string;
  items?: BillItemsItem[];
  amount: number;
  paid: number;
  balance: number;
  status: string;
  /** @nullable */
  paymentMode?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}